training in colab

+++ src/data/ag_vision.py
@@ -98,10 +98,10 @@ class AgVisionSAMDataset(Dataset):
         )
         
         return {
-            "pixel_values": inputs["pixel_values"].squeeze(0),
-            "input_points": inputs["input_points"].squeeze(0),
-            "input_labels": inputs["input_labels"].squeeze(0),
-            "mask": torch.from_numpy(mask).unsqueeze(0) # (1, H, W)
+            "pixel_values": inputs["pixel_values"].squeeze(0).float(),
+            "input_points": inputs["input_points"].squeeze(0).float(),
+            "input_labels": inputs["input_labels"].squeeze(0).float(),
+            "mask": torch.from_numpy(mask).unsqueeze(0).float() # (1, H, W)
         }
 
 class FarmTrackDataModule(pl.LightningDataModule):
